% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{check_FID_IID}
\alias{check_FID_IID}
\title{Check Covariate and Phenotype File First Two Columns and Sample Availability}
\usage{
check_FID_IID(df, iid_chosen)
}
\arguments{
\item{df}{Dataframe of the phenotype or covariates}

\item{iid_chosen}{List of IIDs chosen for analyses}
}
\value{
Dataframe containing the samples with selected IIDs, if available; otherwise, NULL.
}
\description{
This function reads and checks the covariate and phenotype files if they have the 1st and 2nd columns as FID and IID,
and if the input contains less cases than the desired sample, if the input contains the cases for the desired sample
}
