% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{check_input}
\alias{check_input}
\title{Check Input Files}
\usage{
check_input(file_name, file_type)
}
\arguments{
\item{file_name}{A string of the path to the file to be checked for availability}

\item{file_type}{A string of the specific type of file to be checked,
including "SNPs for Extraction", "Participant IDs for Extraction", "Genotype Calls (.pgen/.bed)", "Variant Information (.pvar/.bim)", "Sample Information (.psam/.fam)",
"Output Directory", "Covariate File", "Phenotype File", "GWAS Result File", "Clumped Result File"}
}
\value{
A value 0 or NULL indicating if all criteria are met and procedures can continue.
}
\description{
This function takes the input paths and checks if any paths are missing or unavailable.
}
