% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pipeline.R
\name{check_pipeline}
\alias{check_pipeline}
\title{Test the pipeline of censGWAS() without running PLINK analyses}
\usage{
check_pipeline(
  pfile = NULL,
  bfile = NULL,
  pgen = NULL,
  pvar = NULL,
  psam = NULL,
  bed = NULL,
  bim = NULL,
  fam = NULL,
  file_pheno = NULL,
  file_covar = NULL,
  participants = NULL,
  gwas_thres_hwe = 1e-20,
  gwas_thres_maf = 0.01,
  gwas_thres_geno = 0.05,
  gwas_ci = 0.95,
  gwas_results = NULL,
  snp_id_field = "ID",
  p_field = "P",
  a1_field = "A1",
  a1_freq_field = "A1_FREQ",
  beta_field = "BETA",
  se_field = "SE",
  pos_field = "POS",
  clump_p1 = 5e-08,
  clump_p2 = 1e-04,
  clump_r2 = 0.1,
  clump_kb = 1000,
  file_snps_chosen = NULL,
  output_name = "out",
  delete_unconverted = TRUE,
  standardize_covariates = FALSE
)
}
\arguments{
\item{pfile}{OPTIONAL. String. Name of the PLINK2 genotype files, if .pgen, .pvar, .psam have the same name, no suffix needed in this field. For example, with sample.pgen, sample.pvar, sample.psam, simply provide "sample" for this entry.}

\item{bfile}{OPTIONAL. String. Name of the PLINK1 genotype files, if .bed, .bim, .fam have the same name, no suffix needed in this field. For example, with sample.bed, sample.bim, sample.fam, simply provide "sample" for this entry.}

\item{pgen}{OPTIONAL. A string representing the genotype call file in PLINK2 format, ending with .pgen, should be provided along with pvar and psam in order for the program to proceed.}

\item{pvar}{OPTIONAL. A string representing the variant information file, ending with .pvar.}

\item{psam}{OPTIONAL. A string representing the sample information file, ending with .psam.}

\item{bed}{OPTIONAL. A string representing the genotype call file in PLINK1 format, ending with .bed, should be provided along with bim and fam in order for the program to proceed.}

\item{bim}{OPTIONAL. A string representing the variant information file, ending with .bim.}

\item{fam}{OPTIONAL. A string representing the sample information file, ending with .fam.}

\item{file_pheno}{REQUIRED.A string of the path to the phenotype file (tab- or space-separated) containing the FID, IID, and censored status (0 for below /1 for above LOD), measurements.}

\item{file_covar}{REQUIRED. A string of the path to the covariate file (tab- or space-separated) containing the FID, IID, and covariate columns.}

\item{participants}{REQUIRED. A string of the file including the list of participant IIDs, organized in one column, with header 'IID' (no quote).}

\item{gwas_thres_hwe}{OPTIONAL. Threshold for Hardy-Weignberg equilibrium, default is 1e-20.}

\item{gwas_thres_maf}{OPTIONAL. Threshold for MAF of the analyzed SNPs, default is 0.01.}

\item{gwas_thres_geno}{OPTIONAL. Threshold of missing call rates for filtering out variants. Default is 0.05.}

\item{gwas_ci}{OPTIONAL. Confidence intervals with the given width to be reported for each beta or odds-ratio. Default is 0.95.}

\item{gwas_results}{OPTIONAL. A string of the file including GWAS outcomes, if already performed by the user.}

\item{snp_id_field}{OPTIONAL. The field (column) name in the GWAS result files that contains the SNP IDs, default is 'ID'.}

\item{p_field}{OPTIONAL. The field (column) name in the GWAS result files that contains the p-values, default is 'P'.}

\item{a1_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the effect allele. Default is 'A1'.}

\item{a1_freq_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the A1 allele frequency. Defeult is 'A1_FREQ'.}

\item{beta_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the effect estimate (slope). Default is 'BETA'.}

\item{se_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the standard error. Default is 'SE'.}

\item{pos_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the SNP position. Default is 'POS'.}

\item{clump_p1}{OPTIONAL. The index variant p-value threshold for clumping. Default is 5e-8.}

\item{clump_p2}{OPTIONAL. The SP2 column p-value threshold for clumping. Default is 1e-4.}

\item{clump_r2}{OPTIONAL. The r^2 (correlation between genetic variants to reduce redundancy) threshold for clumping. Default is 0.1.}

\item{clump_kb}{OPTIONAL. The maximum distance from the lead variant for SNPs to be considered to be clumped with. Default is 1000.}

\item{file_snps_chosen}{OPTIONAL.A string representing the file listing the chosen SNPs in one column, if user directly wants to test certain SNPs.}

\item{output_name}{OPTIONAL. A string of the name for the output file, can be the group name or chromosome number.
Default is "out". The corresponding suffices will be added automatically by the program or PLINK2.}

\item{delete_unconverted}{OPTIONAL. Boolean (TRUE or FALSE) to indicate whether remove the extracted but not converted intermediate genotype files or not.
Default is TRUE}

\item{standardize_covariates}{OPTIONAL. Boolean (TRUE or FALSE) to indicate whether the quantitative covariates need to be standardized. Default is FALSE.
PLINK2 may fail if the covariates have not been standardized. When PLINK2 reports the error, turn this flag to be TRUE. Or, provide your already standardized covariates instead.}
}
\value{
A list of steps that has been reached to.
}
\description{
This function tests whether the pipeline of censGWAS() is designed correctly, by taking different input files,
except for the genotype files in PLINK format. To ensure efficiency, GWAS or clumping will not be conducted.
}
