% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{check_snp_list}
\alias{check_snp_list}
\title{Check Format of SNP List}
\usage{
check_snp_list(snp_df)
}
\arguments{
\item{snp_df}{The dataframe read from the text (tab-/space-separated) file containing the SNPs to extract}
}
\value{
Dataframe of the list of selected SNPs, if correctly formatted; otherwise, NULL.
}
\description{
This function checks format of the tab-/space-separated file including SNPs to extract, if directly provided by the user
and returns the converted dataframe if it is correctly formatted with the three columns (Chr., ID, refAllele).
}
