% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_linear_clump.R
\name{clump_gwas}
\alias{clump_gwas}
\title{Clump and organize GWAS Outcomes}
\usage{
clump_gwas(
  gwas_results = NULL,
  p1 = 5e-08,
  p2 = 1e-04,
  r2 = 0.1,
  kb = 1000,
  file_calls = NULL,
  file_variant_info = NULL,
  file_sample_info = NULL,
  plink_file_version,
  output_name = "out",
  snp_id_field = "ID",
  p_field = "P",
  a1_field = "A1",
  a1_freq_field = "A1_FREQ",
  beta_field = "BETA",
  se_field = "SE",
  pos_field = "POS"
)
}
\arguments{
\item{gwas_results}{REQUIRED. A string representing the file containing GWAS results.}

\item{p1}{OPTIONAL. The index variant p-value threshold. Default is 5e-8.}

\item{p2}{OPTIONAL. The SP2 column p-value threshold. Default is 1e-4.}

\item{r2}{OPTIONAL. The r^2 (correlation between genetic variants to reduce redundancy) threshold. Default is 0.1.}

\item{kb}{OPTIONAL. The maximum distance from the lead variant for SNPs to be considered to be clumped with. Default is 1000.}

\item{file_calls}{REQUIRED. A string representing the genotype call file, ending with .pgen or .bed.}

\item{file_variant_info}{REQUIRED. A string representing the variant information files, ending with .pvar or .bim.}

\item{file_sample_info}{REQUIRED. A string representing the sample information file, ending with .psam or .fam.}

\item{plink_file_version}{REQUIRED. Version of PLINK files. Integer 1 or 2. Default is 2.}

\item{output_name}{OPTIONAL. A string of the name for the output file, can be the group name or chromosome number.
Default is "out". The suffix ".clumps" will be added automatically by PLINK.}

\item{snp_id_field}{OPTIONAL. The field (column) name in the GWAS result files that contains the SNP IDs, default is 'ID'.}

\item{p_field}{OPTIONAL. The field (column) name in the GWAS result files that contains the p-values, default is 'P'.}

\item{a1_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the effect allele. Default is 'A1'.}

\item{a1_freq_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the A1 allele frequency. Defeult is 'A1_FREQ'.}

\item{beta_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the effect estimate (slope). Default is 'BETA'.}

\item{se_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the standard error. Default is 'SE'.}

\item{pos_field}{OPTIONAL. The field (column) name in the GWAS result files that contain the SNP position. Default is 'POS'.}
}
\value{
The clumped files (ending with .clumps) and the dataframe containing the clumped SNPs (Chr No., Position, SNP ID, effect allele A1, effect size BETA, standard error, and p-value); or, NULL if process failed.
}
\description{
This function takes the GWAS results (generate by step1_plink() or provided by user) and performs clumping with PLINK2
outputs the dataframe containing the clumped SNPs with chr no., position, SNP ID, effect alleles, effect size, standard error, p-value
}
