% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{get_geno}
\alias{get_geno}
\title{Retrieve Genotype Data}
\usage{
get_geno(
  snps_chosen,
  participants = NULL,
  file_calls = NULL,
  file_variant_info = NULL,
  file_sample_info = NULL,
  plink_file_version,
  output_name = "out",
  delete_unconverted = TRUE
)
}
\arguments{
\item{snps_chosen}{REQUIRED. A dataframe of the SNPs to extract
with 1st column including the chromosome numbers (integer 1~22, string 'X', 'XY'),
2nd column for the SNP IDs (string of rsid),
and 3rd column for the effect alleles (character 'A', 'T', 'G', 'C').}

\item{participants}{REQUIRED. A string representing the file for the FID and IID, listed in two columns, of the participants to extract.}

\item{file_calls}{REQUIRED. A string representing the genotype call file, ending with .pgen or .bed.}

\item{file_variant_info}{REQUIRED. A string representing the variant information files, ending with .pvar or .bim.}

\item{file_sample_info}{REQUIRED. A string representing the sample information file, ending with .psam or .fam.}

\item{plink_file_version}{REQUIRED. Integer indicating the version PLINK, 1 or 2. Generated automatically by identifying the genotype file suffices.}

\item{output_name}{OPTIONAL. A string of the name for the output file, can be the group name or chromosome number.
Default is "out". The suffix is added automatically by PLINK.}

\item{delete_unconverted}{OPTIONAL. Boolean (TRUE or FALSE) to indicate whether remove the extracted but not converted intermediate files or not.
Default is TRUE}
}
\value{
Genotype files: extracted (.bed or .pgen along with the companions) and converted (.raw) for R for subsequent computations,
as well as the a value 0 or NULL to indicate the successfulness of this stage.
}
\description{
This function takes PLINK genotype dosage files (.pgen/.bed, .pvar/.bim, .psam/.fam), participant list,
and table of desired SNPs,
and returns the genotype files (.raw) readable by R.
}
