% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_preparation.R
\name{merge_data}
\alias{merge_data}
\title{Merge Covariate and Phenotype Data Files}
\usage{
merge_data(
  file_covar = NULL,
  file_pheno = NULL,
  participants = NULL,
  output_name = "out",
  need_stdz = FALSE,
  run_linear
)
}
\arguments{
\item{file_covar}{REQUIRED. A string of the path to the covariate file (tab- or space-separated) containing the FID, IID, and covariate columns.}

\item{file_pheno}{REQUIRED. A string of the path to the phenotype file (tab- or space-separated) containing the FID, IID, and censored status (0 for below /1 for above LOD), measurements.}

\item{participants}{REQUIRED. A string representing the file for the IIDs, listed in a column, of the participants to extract.}

\item{output_name}{OPTIONAL. A string of the prefix name to include for the output file, can be the group or strata name.
Default is "out". Function automatically adds suffix ".covar", ".pheno".}

\item{need_stdz}{OPTIONAL. Boolean indicating whether standardization of quantitative covariates is needed. Default is FALSE.}

\item{run_linear}{REQUIRED. Boolean indicating whether to run linear model with PLINK, defined automatically based on if the gwas_results are provided. If TRUE, the phenotype and covariates are outputted as TSV files.}
}
\value{
Dataframe of the covariates and phenotypes (FID, IID, binary censoring status, measurements before and after censoring, rank-transformed values for linear and Tobit, and the covariates), with complete cases,
along with the corresponding tsv files for PLINK use for running linear model, if chosen to run.
If the procedures were not able to continue, returns NULL.
}
\description{
This function reads the covariate and phenotype files and participant ID
and returns the organized dataframe after checking for complete cases
}
