% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_HWE_geno}
\alias{simulate_HWE_geno}
\title{Simulate the Genotypes based on Hardy-Weiberg Equilibrium}
\usage{
simulate_HWE_geno(FIDs, IIDs, mafs, N)
}
\arguments{
\item{FIDs}{A list of simulated FIDs.}

\item{IIDs}{A list of simulated IIDs.}

\item{mafs}{A list floating-point numeric values less than 0.5 indicating the minor allele frequency of the SNPs.}

\item{N}{An integer indicating the sample size.}
}
\value{
A dataframe containing the FID, IID, dummy PAT, dummy MAT, dummy SEX, dummy PHENOTYPE, and the SNP with dummy RS-ID.
}
\description{
This function takes the MAF and sample size to generate genotype data for testing.
}
