% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulation.R
\name{simulate_phenotype}
\alias{simulate_phenotype}
\title{Simulate Phenotype using the Genotypes and their Corresponding Effect Sizes}
\usage{
simulate_phenotype(genotypes, betas, covar, cens_prop, to_truncate)
}
\arguments{
\item{genotypes}{The dataframe generated by simulate_HWE_geno().}

\item{betas}{The effect sizes of each SNP}

\item{covar}{The covariate, in the simulation here we have only 1 covariate for simplicity}

\item{cens_prop}{The proportion of measurements that should be below the LOD.}

\item{to_truncate}{Boolean indicating whether to have truncated values (NA) or censored. Just for testing the function merge_data().}
}
\value{
A dataframe containing the FID, IID, censored status, and measured values.
}
\description{
Simulate Phenotype using the Genotypes and their Corresponding Effect Sizes
}
