% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_linear_clump.R
\name{snps_to_extract}
\alias{snps_to_extract}
\title{List SNPs for Extraction from the Clumped Results}
\usage{
snps_to_extract(clumped_df)
}
\arguments{
\item{clumped_df}{REQUIRED. The dataframe of the clumped results, generated by clump_gwas()}
}
\value{
dataframe containing chromosome number, SNP ID, A1
}
\description{
This function gets the Chromosome Number, SNP ID, and the effect allele A1 from the clumped dataframe generated by organize_clumped() or
and generates the dataframe containing the three columns described for get_geno()
}
