% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step2_refinement_tobit.R
\name{step2_run_tobit}
\alias{step2_run_tobit}
\title{Refine Estimates with Tobit Model}
\usage{
step2_run_tobit(
  geno_converted = NULL,
  pheno_covar = NULL,
  snps_to_refine = NULL,
  output_name = "out"
)
}
\arguments{
\item{geno_converted}{REQUIRED. A string representing the converted genotype file (.raw) generated by get_geno().}

\item{pheno_covar}{REQUIRED. The datafame containing the FID, IID, phenotypes and covariates, organized by merge_data().}

\item{snps_to_refine}{REQUIRED. The dataframe containing the clumped results or selected snps for refinement, clumping generated by either step1_plink(), or SNPs provided by users themselves.}

\item{output_name}{OPTIONAL. A string for the output file suffix name containing the refined result. Default is "out"}
}
\value{
A text file including the refined estimates of the SNPs using Tobit model.
}
\description{
This function takes the converted genotype, a dataframe of covariate and phenotype,
and the table of snps selected either as the clumping results or three-column table containing only the chromosome number, SNP ID and effect allele
and gives the refined effect estimates (intercept, BETA and SE, statistic, p-value) with Tobit model in textfile,
}
